get stars from storage into maplevel
